Extract raw api to be implemented in Project1 into raw_api.go
