+++ kv/server/raw_api.go
@@ -1,0 +1,29 @@
+package server
+
+import (
+	"context"
+	"github.com/pingcap-incubator/tinykv/proto/pkg/kvrpcpb"
+)
+
+// The functions below are Server's Raw API. (implements TinyKvServer).
+// Some helper methods can be found in sever.go in the current directory
+
+func (server *Server) RawGet(_ context.Context, req *kvrpcpb.RawGetRequest) (*kvrpcpb.RawGetResponse, error) {
+	// Your Code Here (1).
+	return nil, nil
+}
+
+func (server *Server) RawPut(_ context.Context, req *kvrpcpb.RawPutRequest) (*kvrpcpb.RawPutResponse, error) {
+	// Your Code Here (1).
+	return nil, nil
+}
+
+func (server *Server) RawDelete(_ context.Context, req *kvrpcpb.RawDeleteRequest) (*kvrpcpb.RawDeleteResponse, error) {
+	// Your Code Here (1).
+	return nil, nil
+}
+
+func (server *Server) RawScan(_ context.Context, req *kvrpcpb.RawScanRequest) (*kvrpcpb.RawScanResponse, error) {
+	// Your Code Here (1).
+	return nil, nil
+}
